v0.0.3
On token control event, pushes the token to the top of the viewed stack for all connected clients who are viewing the same scene.
All you have to do is install and enable this module. There are no user interface elements.
If a player can only control one token (their PC), they won't be able to bring their token to the top of the stack by clicking on it as no controlToken hook is fired, but this will work if they deselect their token first.